Low: crmd: Move Alert function to common.
