+++ crmd/notify.c
@@ -22,219 +22,15 @@
 #include <crm/pengine/rules.h>
 #include "notify.h"
 #include "crmd_messages.h"
-
-static char *notify_script = NULL;
-static char *notify_target = NULL;
-static GListPtr notify_list = NULL;
-static int alerts_inflight = 0;
-static gboolean draining_alerts = FALSE;
-static guint max_alert_timeout = CRMD_NOTIFY_DEFAULT_TIMEOUT_MS;
-
-typedef struct {
-    char *name;
-    char *value;
-} envvar_t;
-
-typedef struct {
-    char *id;
-    char *path;
-    int timeout;
-    char *tstamp_format;
-    char *recipient;
-    GListPtr envvars;
-} notify_entry_t;
-
-enum notify_keys_e{
-    CRM_notify_recipient = 0,
-    CRM_notify_node,
-    CRM_notify_nodeid,
-    CRM_notify_rsc,
-    CRM_notify_task,
-    CRM_notify_interval,
-    CRM_notify_desc,
-    CRM_notify_status,
-    CRM_notify_target_rc,
-    CRM_notify_rc,
-    CRM_notify_kind,
-    CRM_notify_version,
-    CRM_notify_node_sequence,
-    CRM_notify_timestamp
-};
-
-/*
- * to allow script compatibility we can have more than one
- * set of environment variables
- */
-
-static const char *notify_keys[][3] =
-{
-    [CRM_notify_recipient]     = {"CRM_notify_recipient",     "CRM_alert_recipient",     NULL},
-    [CRM_notify_node]          = {"CRM_notify_node",          "CRM_alert_node",          NULL},
-    [CRM_notify_nodeid]        = {"CRM_notify_nodeid",        "CRM_alert_nodeid",        NULL},
-    [CRM_notify_rsc]           = {"CRM_notify_rsc",           "CRM_alert_rsc",           NULL},
-    [CRM_notify_task]          = {"CRM_notify_task",          "CRM_alert_task",          NULL},
-    [CRM_notify_interval]      = {"CRM_notify_interval",      "CRM_alert_interval",      NULL},
-    [CRM_notify_desc]          = {"CRM_notify_desc",          "CRM_alert_desc",          NULL},
-    [CRM_notify_status]        = {"CRM_notify_status",        "CRM_alert_status",        NULL},
-    [CRM_notify_target_rc]     = {"CRM_notify_target_rc",     "CRM_alert_target_rc",     NULL},
-    [CRM_notify_rc]            = {"CRM_notify_rc",            "CRM_alert_rc",            NULL},
-    [CRM_notify_kind]          = {"CRM_notify_kind",          "CRM_alert_kind",          NULL},
-    [CRM_notify_version]       = {"CRM_notify_version",       "CRM_alert_version",       NULL},
-    [CRM_notify_node_sequence] = {"CRM_notify_node_sequence", "CRM_alert_node_sequence", NULL},
-    [CRM_notify_timestamp]     = {"CRM_notify_timestamp",     "CRM_alert_timestamp",     NULL}
-};
-
+#include <crm/common/alerts_internal.h>
 #include <crm/common/iso8601_internal.h>
 
-/*
- * end of possibly generic time-handling stuff
- */
-
+static int alerts_inflight = 0;
+static gboolean draining_alerts = FALSE;
 
 /*
  * syncronize local data with cib
  */
-
-static void
-free_envvar_entry(envvar_t *entry)
-{
-    free(entry->name);
-    free(entry->value);
-    free(entry);
-}
-
-static void
-free_notify_list_entry(notify_entry_t *entry)
-{
-    free(entry->id);
-    free(entry->path);
-    free(entry->tstamp_format);
-    free(entry->recipient);
-    if (entry->envvars) {
-        g_list_free_full(entry->envvars,
-                         (GDestroyNotify) free_envvar_entry);
-    }
-    free(entry);
-}
-
-static void
-free_notify_list()
-{
-    if (notify_list) {
-        g_list_free_full(notify_list, (GDestroyNotify) free_notify_list_entry);
-        notify_list = NULL;
-    }
-}
-
-static gpointer
-copy_envvar_entry(envvar_t * src,
-                  gpointer data)
-{
-    envvar_t *dst = calloc(1, sizeof(envvar_t));
-
-    CRM_ASSERT(dst);
-    dst->name = strdup(src->name);
-    dst->value = src->value?strdup(src->value):NULL;
-    return (gpointer) dst;
-}
-
-static GListPtr
-add_dup_envvar(GListPtr envvar_list,
-               envvar_t *entry)
-{
-    return g_list_prepend(envvar_list, copy_envvar_entry(entry, NULL));
-}
-
-static GListPtr
-drop_envvars(GListPtr envvar_list, int count)
-{
-    int i;
-
-    for (i = 0;
-         (envvar_list) && ((count < 0) || (i < count));
-         i++) {
-        free_envvar_entry((envvar_t *) g_list_first(envvar_list)->data);
-        envvar_list = g_list_delete_link(envvar_list,
-                                         g_list_first(envvar_list));
-    }
-    return envvar_list;
-}
-
-static GListPtr
-copy_envvar_list_remove_dupes(GListPtr src)
-{
-    GListPtr dst = NULL, ls, ld;
-
-    /* we are adding to the front so variable dupes coming via
-     * recipient-section have got precedence over those in the
-     * global section - we don't expect that many variables here
-     * that it pays off to go for a hash-table to make dupe elimination
-     * more efficient - maybe later when we might decide to do more
-     * with the variables than cycling through them
-     */
-
-    for (ls = g_list_first(src); ls; ls = g_list_next(ls)) {
-        for (ld = g_list_first(dst); ld; ld = g_list_next(ld)) {
-            if (!strcmp(((envvar_t *)(ls->data))->name,
-                        ((envvar_t *)(ld->data))->name)) {
-                break;
-            }
-        }
-        if (!ld) {
-            dst = g_list_prepend(dst,
-                    copy_envvar_entry((envvar_t *)(ls->data), NULL));
-        }
-    }
-
-    return dst;
-}
-
-static void
-add_dup_notify_list_entry(notify_entry_t *entry)
-{
-    notify_entry_t *new_entry =
-        (notify_entry_t *) calloc(1, sizeof(notify_entry_t));
-
-    CRM_ASSERT(new_entry);
-    *new_entry = (notify_entry_t) {
-        .id = strdup(entry->id),
-        .path = strdup(entry->path),
-        .timeout = entry->timeout,
-        .tstamp_format = entry->tstamp_format?strdup(entry->tstamp_format):NULL,
-        .recipient = entry->recipient?strdup(entry->recipient):NULL,
-        .envvars = entry->envvars?
-            copy_envvar_list_remove_dupes(entry->envvars)
-            :NULL
-    };
-    notify_list = g_list_prepend(notify_list, new_entry);
-}
-
-static GListPtr
-get_envvars_from_cib(xmlNode *basenode, GListPtr list, int *count)
-{
-    xmlNode *envvar;
-    xmlNode *pair;
-
-    if ((!basenode) ||
-        (!(envvar = first_named_child(basenode, XML_TAG_ATTR_SETS)))) {
-        return list;
-    }
-
-    for (pair = first_named_child(envvar, XML_CIB_TAG_NVPAIR);
-         pair; pair = __xml_next(pair)) {
-
-        envvar_t envvar_entry = (envvar_t) {
-            .name = (char *) crm_element_value(pair, XML_NVPAIR_ATTR_NAME),
-            .value = (char *) crm_element_value(pair, XML_NVPAIR_ATTR_VALUE)
-        };
-        crm_trace("Found environment variable %s = '%s'", envvar_entry.name,
-                  envvar_entry.value?envvar_entry.value:"");
-        (*count)++;
-        list = add_dup_envvar(list, &envvar_entry);
-    }
-
-    return list;
-}
 
 static GHashTable *
 get_meta_attrs_from_cib(xmlNode *basenode, notify_entry_t *entry,
